Remove d0fdcb04 incorrect link. Change default link to browsing page. Refactor LoginForm and SignupForm to Auth/**.
