+++ frontend/src/App.js
@@ -76,14 +76,6 @@ const App = () => {
         </RequireAuth>
       ),
     },
-    {
-      path: "/ranking",
-      element: (
-        <RequireAuth>
-          <Ranking />
-        </RequireAuth>
-      ),
-    },
   ]);
 
   return <RouterProvider router={router} />;
