+++ frontend/src/components/Navbar.js
@@ -1,7 +1,8 @@
-import React, { useState, useContext } from "react";
+import { useState, useContext } from "react";
 import { Transition } from "@headlessui/react";
 import { Link } from "react-router-dom";
 
+import { MdMenuOpen, MdClose } from "react-icons/md/index.js";
 import { Context } from "../context.js";
 
 const Actions = ({ auth, customStyle }) => {
@@ -9,15 +10,15 @@ const Actions = ({ auth, customStyle }) => {
     <>
       {auth ? (
         <>
-          <Link to="/ranking" className={customStyle}>
-            Ranking
+          <Link to="/home" className={customStyle}>
+            Home
           </Link>
           <Link to="/leaveareview" className={customStyle}>
             Leave Review
           </Link>
-          <button type="button" className={customStyle}>
+          <Link to="/logout" className={customStyle}>
             Logout
-          </button>
+          </Link>
         </>
       ) : (
         <>
@@ -33,9 +34,21 @@ const Actions = ({ auth, customStyle }) => {
   );
 };
 
-const Navbar = ({ auth }) => {
+const Navbar = () => {
   const { state } = useContext(Context);
   const [isOpen, setIsOpen] = useState(false);
+
+  const mainstyle =
+    "hover:bg-gray-700 text-white block px-3 py-2 rounded-md text-base font-medium";
+
+  const sidstyle =
+    "hover:bg-gray-700 text-white px-3 py-2 rounded-md text-sm font-medium";
+
+  const mobileStyle =
+    "bg-gray-900 inline-flex items-center justify-center p-2 rounded-md text-gray-400 hover:text-white hover:bg-gray-800 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-offset-gray-800 focus:ring-white";
+  
+  
+
   return (
     <>
       <nav className="bg-gray-800 fixed w-full z-20 top-0 left-0">
@@ -44,63 +57,25 @@ const Navbar = ({ auth }) => {
             <div className="flex items-center">
               <div className="hidden md:block">
                 <div className="ml-10 flex items-baseline space-x-4">
-                  <Link
-                    to="/home"
-                    className=" hover:bg-gray-700 text-white px-3 py-2 rounded-md text-sm font-medium"
-                  >
-                    Home
+                  <Link to="/" className={mainstyle}>
+                    Browsing
                   </Link>
                 </div>
               </div>
             </div>
             <div className="hidden md:block">
-              <Actions
-                auth={state.auth}
-                customStyle="hover:bg-gray-700 text-white px-3 py-2 rounded-md text-sm font-medium"
-              />
+              <Actions auth={state.auth} customStyle={sidstyle} />
             </div>
             <div className="-mr-2 flex md:hidden">
               <button
                 onClick={() => setIsOpen(!isOpen)}
                 type="button"
-                className="bg-gray-900 inline-flex items-center justify-center p-2 rounded-md text-gray-400 hover:text-white hover:bg-gray-800 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-offset-gray-800 focus:ring-white"
+                className={mobileStyle}
                 aria-controls="mobile-menu"
                 aria-expanded="false"
               >
                 <span className="sr-only">Open main menu</span>
-                {!isOpen ? (
-                  <svg
-                    className="block h-6 w-6"
-                    xmlns="http://www.w3.org/2000/svg"
-                    fill="none"
-                    viewBox="0 0 24 24"
-                    stroke="currentColor"
-                    aria-hidden="true"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth="2"
-                      d="M4 6h16M4 12h16M4 18h16"
-                    />
-                  </svg>
-                ) : (
-                  <svg
-                    className="block h-6 w-6"
-                    xmlns="http://www.w3.org/2000/svg"
-                    fill="none"
-                    viewBox="0 0 24 24"
-                    stroke="currentColor"
-                    aria-hidden="true"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth="2"
-                      d="M6 18L18 6M6 6l12 12"
-                    />
-                  </svg>
-                )}
+                {!isOpen ? <MdMenuOpen /> : <MdClose />}
               </button>
             </div>
           </div>
@@ -118,16 +93,10 @@ const Navbar = ({ auth }) => {
           {(ref) => (
             <div className="md:hidden" id="mobile-menu">
               <div ref={ref} className="px-2 pt-2 pb-3 space-y-1 sm:px-3">
-                <Link
-                  to="/home"
-                  className="hover:bg-gray-700 text-white block px-3 py-2 rounded-md text-base font-medium"
-                >
-                  Home
+                <Link to="/" className={mainstyle}>
+                  Ranking
                 </Link>
-                <Actions
-                  auth={state.auth}
-                  customStyle="hover:bg-gray-700 text-white block px-3 py-2 rounded-md text-base font-medium"
-                />
+                <Actions auth={state.auth} customStyle={mainstyle} />
               </div>
             </div>
           )}
