+++ frontend/src/pages/Login.js
@@ -1,5 +1,5 @@
 import React from "react";
-import LoginForm from "../components/LoginForm.js";
+import LoginForm from "../components/Auth/LoginForm.js";
 import logo from "../assets/logo.png";
 import BasePage from "./BasePage.js";
 
@@ -10,10 +10,10 @@ const Login = () => {
         <div className="min-h-full flex items-center justify-center py-12 px-4 sm:px-6 lg:px-8">
           <div className="max-w-md w-full space-y-8">
             <div>
-              <img className="mx-auto h-12 w-auto" src={logo} alt="Workflow" />
-              <h2 className="mt-6 text-center text-3xl font-extrabold text-gray-900">
+              <img className="mx-auto h-12 w-auto" src={logo} alt="Logo" />
+              <h1 className="mt-6 text-center text-3xl font-extrabold text-gray-900">
                 Log in to your account
-              </h2>
+              </h1>
             </div>
             <LoginForm />
           </div>
